+++ Jira/Workflows/process-to-review.js
@@ -4,8 +4,9 @@
 // conditions: true
 
 def summary = issue.fields['summary'] as String
+def ticketType = issue.fields.issuetype?.name?.toLowerCase()
 if(summary.toLowerCase().trim() == "test") {
-    logger.info("Ignore test ${issue.fields.issuetype.name.toLowerCase()} ${issue.key}")
+    logger.info("Ignore test ${ticketType} ${issue.key}")
     return
 }
 
@@ -26,256 +27,11 @@ if(null == processCode) {
 
 def processOwnerId = customFields.find { it.name == 'Process owner' }?.id?.toString()
 def processManagerId = customFields.find { it.name == 'Process manager' }?.id?.toString()
-def processHomeId = customFields.find { it.name == 'Process homepage' }?.id?.toString()
-def processReviewGuideId = customFields.find { it.name == 'Process review guide' }?.id?.toString()
-def definitionUpdatesId = customFields.find { it.name == 'Process definition review and updates' }?.id?.toString()
-def policyHomeId = customFields.find { it.name == 'Policy homepage' }?.id?.toString()
-def procedureHomeId = customFields.find { it.name == 'Procedure homepage' }?.id?.toString()
-def policyUpdatesId = customFields.find { it.name == 'Policy review and updates' }?.id?.toString()
-def procedureUpdatesId = customFields.find { it.name == 'Procedure review and updates' }?.id?.toString()
-def kpiUpdatesId = customFields.find { it.name == 'Performance indicator review and updates' }?.id?.toString()
-def reportUpdatesId = customFields.find { it.name == 'Report review and updates' }?.id?.toString()
 def reviewFrequencyId = customFields.find { it.name == 'Review process' }?.id?.toString()
 
 def processOwner = issue.fields[processOwnerId]?.accountId as String
 def processManager = issue.fields[processManagerId]?.accountId as String
-def processHome = issue.fields[processHomeId] as String
-def processReviewGuide = issue.fields[processReviewGuideId]
 def reviewFrequency = issue.fields[reviewFrequencyId]?.value as String
-
-def definitionUpdates = [
-    type: "doc",
-    version: 1,
-    content: [
-        [
-            type: "heading",
-            attrs: [ level: 3 ],
-            content: [[
-                type: "text",
-                text: "Goals",
-            ]]
-        ],
-        [
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements.",
-            ]]
-        ],
-        [
-            type: "heading",
-            attrs: [ level: 3 ],
-            content: [[
-                type: "text",
-                text: "Requirements",
-            ]]
-        ],
-        [
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements.",
-            ]]
-        ],
-        [
-            type: "heading",
-            attrs: [ level: 3 ],
-            content: [[
-                type: "text",
-                text: "Roles",
-            ]]
-        ],
-        [
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements.",
-            ]]
-        ],
-        [
-            type: "heading",
-            attrs: [ level: 3 ],
-            content: [[
-                type: "text",
-                text: "Input & Output",
-            ]]
-        ],
-        [
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements.",
-            ]]
-        ],
-    ]
-]
-
-def policyUpdates = [
-    type: "doc",
-    version: 1,
-    content: []
-]
-
-def procedureUpdates = [
-    type: "doc",
-    version: 1,
-    content: []
-]
-
-def kpiUpdates = [
-    type: "doc",
-    version: 1,
-    content: []
-]
-
-def reportUpdates = [
-    type: "doc",
-    version: 1,
-    content: [
-        [
-            type: "heading",
-            attrs: [ level: 2 ],
-            content: [[
-                type: "text",
-                text: "Report Type",
-            ]]
-        ],
-        [
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements. (Repeat as needed)",
-            ]]
-        ],
-    ]
-]
-
-logger.info(processHome)
-
-if(null != processHome) {
-    definitionUpdates.content[0]?.content[0]?.marks = [[
-        type: "link",
-        attrs: [ href: "${processHome}#Goals" ]
-    ]]
-
-    definitionUpdates.content[2]?.content[0]?.marks = [[
-        type: "link",
-        attrs: [ href: "${processHome}#Requirements" ]
-    ]]
-
-    definitionUpdates.content[4]?.content[0]?.marks = [[
-        type: "link",
-        attrs: [ href: "${processHome}#Roles" ]
-    ]]
-
-    definitionUpdates.content[6]?.content[0]?.marks = [[
-        type: "link",
-        attrs: [ href: "${processHome}#Inputs-%26-Outputs" ]
-    ]]
-}
-
-// find all active policies
-def result = get("/rest/api/3/search/jql?fields=key,summary,status,${policyHomeId}&jql=project%3D${processCode}%20AND%20issuetype%3DPolicy%20AND%20status%20!%3D%20Inactive")
-    .asObject(Map)
-
-if(result.status >= 200 && result.status < 300) {
-    for(def pol : result.body.issues) {
-
-        def heading = [
-            type: "heading",
-            attrs: [ level: 3 ],
-            content: [[
-                type: "text",
-                text: pol?.fields?.summary,
-            ]]
-        ]
-
-        if(null != pol?.fields[policyHomeId])
-            heading.content[0].marks = [[
-                type: "link",
-                attrs: [ href: pol?.fields[policyHomeId] ]
-            ]]
-        
-        policyUpdates.content.add(heading)
-        policyUpdates.content.add([
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements.",
-            ]]
-        ])
-    }
-}
-else
-    logger.info("Could not list ${processCode} policies (${result.status})")
-
-// find all active procedures
-result = get("/rest/api/3/search/jql?fields=key,summary,status,${procedureHomeId}&jql=project%3D${processCode}%20AND%20issuetype%3DProcedure%20AND%20status%20!%3D%20Inactive")
-    .asObject(Map)
-
-if(result.status >= 200 && result.status < 300) {
-    for(def proc : result.body.issues) {
-
-        def heading = [
-            type: "heading",
-            attrs: [ level: 3 ],
-            content: [[
-                type: "text",
-                text: proc?.fields?.summary,
-            ]]
-        ]
-
-        if(null != proc?.fields[procedureHomeId])
-            heading.content[0].marks = [[
-                type: "link",
-                attrs: [ href: proc?.fields[procedureHomeId] ]
-            ]]
-        
-        procedureUpdates.content.add(heading)
-        procedureUpdates.content.add([
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements.",
-            ]]
-        ])
-    }
-}
-else
-    logger.info("Could not list ${processCode} procedures (${result.status})")
-
-// find all active KPIs
-result = get("/rest/api/3/search/jql?fields=key,summary,status&jql=project%3D${processCode}%20and%20issuetype%3D%22Key%20Performance%20Indicator%22%20and%20status!%3DInactive")
-    .asObject(Map)
-
-if(result.status >= 200 && result.status < 300) {
-    for(def kpi : result.body.issues) {
-
-        def heading = [
-            type: "heading",
-            attrs: [ level: 3 ],
-            content: [[
-                type: "text",
-                text: kpi.fields?.summary,
-                marks: [[
-                    type: "link",
-                    attrs: [ href: "/browse/${kpi?.key}" ]
-                ]]
-            ]]
-        ]
-
-        kpiUpdates.content.add(heading)
-        kpiUpdates.content.add([
-            type: "paragraph",
-            content: [[
-                type: "text",
-                text: "Current status and need for improvements.",
-            ]]
-        ])
-    }
-}
-else
-    logger.info("Could not list ${processCode} KPIs (${result.status})")
 
 // create Process Review ticket in the correct Jira project
 def assignee = null
@@ -324,20 +80,6 @@ result = post("/rest/api/3/issue")
             issuetype: [ name: "Process Review" ],
             summary: "Review of process ${processCode} on ${reviewDate}",
             assignee: assignee,
-            description: processReviewGuide,
-            (definitionUpdatesId): definitionUpdates,
-            (policyUpdatesId): policyUpdates,
-            (procedureUpdatesId): procedureUpdates,
-            (kpiUpdatesId): kpiUpdates,
-            (reportUpdatesId): reportUpdates,
-        ],
-        update:[
-            issuelinks: [[
-                add: [
-                    type: [ name: "Review" ],
-                    inwardIssue: [ key: issue.key ]
-                ]
-            ]]
         ],
     ])
     .asObject(Map)
@@ -349,19 +91,6 @@ if(result.status < 200 || result.status >= 300) {
 
 def newTicket = result.body as Map
 logger.info("Created process review ${newTicket.key}")
-
-// assign to process owner
-result = put("/rest/api/3/issue/${issue.key}")
-    .header("Content-Type", "application/json")
-    .body([
-        fields: [
-            assignee: assignee,
-        ],
-    ])
-    .asString()
-
-if(result.status < 200 || result.status > 204)
-    logger.info("Could not update process ${issue.key} (${result.status})")
 
 // add a comment about the new review that has started
 result = post("/rest/api/3/issue/${issue.key}/comment")